requirements.txt
alembic==0.8.5
appdirs==1.4.0
Flask==0.10.1
Flask-Cors==3.0.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
GoogleAppEngineCloudStorageClient==1.9.0.0
itsdangerous==0.24
Jinja2==2.8
PyJWT==1.4.2
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
packaging==16.8
pycrypto==2.6.1
pyparsing==2.1.10
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.10
